Navigation Part 1
Installed the react-navigation library and learned to use it to create and configure a StackNavigator for the Directory and CampsiteInfo screens.
Learned about react-navigation's navigation prop, and how to use it to control navigating from the Directory to a selected CampsiteInfo component using params.
Learned a typical way to separate code for iOS and Android platforms.